🚚️ api: move repository to src/enrolment context
