+++ api/src/certification/enrolment/infrastructure/repositories/index.js
@@ -8,6 +8,7 @@ import * as certificationCandidateRepository from './certification-candidate-rep
 import * as certificationCpfCityRepository from './certification-cpf-city-repository.js';
 import * as certificationCpfCountryRepository from './certification-cpf-country-repository.js';
 import * as complementaryCertificationRepository from './complementary-certification-repository.js';
+import * as countryRepository from './country-repository.js';
 import * as sessionForAttendanceSheetRepository from './session-for-attendance-sheet-repository.js';
 import * as sessionRepository from './session-repository.js';
 
@@ -24,6 +25,7 @@ import * as sessionRepository from './session-repository.js';
  * @typedef {certificationCpfCityRepository} CertificationCpfCityRepository
  * @typedef {sessionForAttendanceSheetRepository} SessionForAttendanceSheetRepository
  * @typedef {sessionManagementRepository} SessionManagementRepository
+ * @typedef {countryRepository} CountryRepository
  */
 const repositoriesWithoutInjectedDependencies = {
   candidateRepository,
@@ -36,6 +38,7 @@ const repositoriesWithoutInjectedDependencies = {
   certificationCpfCityRepository,
   sessionForAttendanceSheetRepository,
   sessionManagementRepository,
+  countryRepository,
 };
 
 /**
